authorization checks

+++ controllers/feed.js
@@ -138,6 +138,12 @@ exports.updatePost = (req, res, next) => {
                 throw error
             }
 
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized')
+                error.statusCode = 403
+                throw error
+            }
+
             if (imageUrl !== post.imageUrl) {
                 clearImage(post.imageUrl)
             }
@@ -170,6 +176,13 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404
                 throw error
             }
+
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized')
+                error.statusCode = 403
+                throw error
+            }
+
             // TODO check if post author is current user and can delete
             clearImage(post.imageUrl)
             return Post.findByIdAndRemove(id)
